split soragodong code
