+++ routes/index.js
@@ -4,6 +4,8 @@ const router = express.Router();
 
 const libKakaoWork = require('../libs/kakaoWork');
 
+const sora = require('../sora');
+
 router.get('/', async (req, res, next) => {
 	// 유저 목록 검색 (1)
 	const users = await libKakaoWork.getUserList();
@@ -65,7 +67,6 @@ router.get('/', async (req, res, next) => {
 		),
 	]);
 
-	// 응답값은 자유롭게 작성하셔도 됩니다.
 	res.json({
 		users,
 		conversations,
@@ -73,34 +74,12 @@ router.get('/', async (req, res, next) => {
 	});
 });
 
-// routes/index.js
 router.post('/request', async (req, res, next) => {
 	const { message, value } = req.body;
 
 	switch (value) {
 		case 'ask_sora':
-			// 설문조사용 모달 전송
-			return res.json({
-				view: {
-					title: '마법의 소라고동',
-					accept: '확인',
-					decline: '취소',
-					value: 'question_for_sora',
-					blocks: [
-						{
-							type: 'label',
-							text: '*_마법 고동님, 물어볼 게 있는데요,_*',
-							markdown: true,
-						},
-						{
-							type: 'input',
-							name: 'question',
-							required: false,
-							placeholder: '질문을 입력해주세요.',
-						},
-					],
-				},
-			});
+			return res.json(sora.ask_sora_modal);
 			break;
 		default:
 	}
@@ -108,65 +87,16 @@ router.post('/request', async (req, res, next) => {
 	res.json({});
 });
 
-// routes/index.js
 router.post('/callback', async (req, res, next) => {
-	const { message, actions, action_time, value } = req.body; // 설문조사 결과 확인 (2)
+	const { message, actions, action_time, value } = req.body;
 
 	switch (value) {
 		case 'question_for_sora':
-			let answer = [
-				'*_안 돼._*',
-				'*_가만히 있어._*',
-				'*_멈춰._*',
-				'*_그럼._*',
-				'*_다시 한번 물어봐._*',
-			];
-			let picture = [
-				'https://user-images.githubusercontent.com/45932570/115988069-eeae4880-a5f2-11eb-8f6a-b4ae24311e94.jpg',
-				'https://user-images.githubusercontent.com/45932570/115988103-1c938d00-a5f3-11eb-8231-b358115c061d.png',
-				'https://user-images.githubusercontent.com/45932570/115987785-7dba6100-a5f1-11eb-996a-656c77a1d956.jpg',
-				'https://user-images.githubusercontent.com/45932570/115988194-77c57f80-a5f3-11eb-8145-78611e1663c5.png',
-				'https://user-images.githubusercontent.com/45932570/115988194-77c57f80-a5f3-11eb-8145-78611e1663c5.png',
-			];
-			var no = Math.floor(Math.random() * 10) % 4;
-
-			// 설문조사 응답 결과 메세지 전송 (3)
-			await libKakaoWork.sendMessage({
-				conversationId: message.conversation_id,
-				text: 'Push alarm message',
-				blocks: [
-					{
-						type: 'image_link',
-						url: picture[no],
-					},
-					{
-						type: 'text',
-						text: answer[no],
-						markdown: true,
-					},
-				],
-			});
+			await libKakaoWork.sendMessage(sora.from_sora(message.conversation_id));
 			break;
 			
 		case 'start_sora':
-			console.log("start_sora : conversation_id: " + message.conversation_id);
-			await libKakaoWork.sendMessage({
-				conversationId: message.conversation_id,
-				text: '마법의 소라고동',
-				blocks: [
-					{
-					  "type": "image_link",
-					  "url": "https://user-images.githubusercontent.com/45932570/115987984-83fd0d00-a5f2-11eb-8a54-d6ee9d0f8085.jpg"
-					},
-					{
-						type: 'button',
-						action_type: 'call_modal',
-						value: 'ask_sora',
-						text: '소라고동에게 물어보기',
-						style: 'default',
-					},
-				],
-			});
+			await libKakaoWork.sendMessage(sora.start_sora(message.conversation_id));
 			break;
 		default:
 	}
